# feat(python): add LRU cache for parsed contract classes