According to Starkware a large amount of time is spent on de-serializing classes from their JSON representation. To make things faster they've added a cache for the parsed class instances.

This PR adds a shared LRU cache that caches the parsed classes between commands -- we hope to have better performance this way. This should be especially noticeable for quick calls.